[lio-py]: Add missing lio_node.py --addtpg op
This patch adds the lio_node.py --addtpg op for completeness with the
already existing --deltpg.  Note that is most cases lio_node.py --addnp,
--addlun or --addnodeacl will automatically create the tpg endpoint when
calling one of these three operations because of the recursive nature of
configfs directory group creation, which means that --addtpg is not required
when --addnp, --addlun or --addnodeacl is called.

Signed-off-by: Nicholas A. Bellinger <nab@linux-iscsi.org>