[tcm-py]: Add tcm_node.py --scsibyudev op for Linux/SCSI devices
This patch adds a tcm_node.py --scsibyudev op that allows for
TCM/pSCSI to extract HCTL from a given udev path using /sys/block to
locate a Linux/SCSI device:

  --scsibyudev=DEV <UDEV_PATH>, --pscsibyudev=DEV <UDEV_PATH>
                        Associate TCM/pSCSI object with Linux/SCSI device by
                        UDEV Path

Note that the first parameter for --scsibyudev is 'DEV' (and not 'HBA/DEV' as for
--scsi, --block, etc) and will automatically register the pSCSI HBA based on the
Linux/SCSI Host ID, and set the internal udev_path presented by configfs.

Note that tcm_dump.py does not support udev_path for pSCSI yet, and will output
the HCTL parameters for TCM/pSCSI storage objects in the same manner for
tcm_node.py --scsibyudev as well as --scsi.

Signed-off-by: Nicholas A. Bellinger <nab@linux-iscsi.org>