tcm_node: Use helper fn to let all callbacks take desired args only
Instead of making all callbacks take the four arguments defined by
OptParse, with the real arguments in the value list, we pass the
actual callback function to a dispatch function, which unpacks the
value list and calls the original callback with the values as args.

This should help readability overall.

Signed-off-by: Andy Grover <agrover@redhat.com>
Signed-off-by: Nicholas A. Bellinger <nab@linux-iscsi.org>